add support for local address selection
when testing with large connection number on fast links,
the local port selection becomes the test bottle-neck, because
the local TCP sockets are hashed to the same entry into the
kernel bind table as all of them use the same destination ip
and same destination port.

Using many different local ip address improve a lot the connection
rate. This change introduce support for 2 new command line arguments:
'local_ip', and 'range'.

When the first is specified, the created connection will bind to
that address. If range is specified, too, the created connection will
bind to the address in range <local_ip>..<local_ip + range> using
a round robin policy. Is up to the caller setup the addresses
accordingly.